mv2_cortex_m_mve: pack per-channel requant shifts as int8
Shifts come from frexp(scale) — the exponent for a float scale that's
always small (cortex_m pipeline emits shifts in roughly [-30, +5];
observed [-13, -6] across all 17,056 channels of MV2).  They were
being stored as int32 arrays solely so the MVE per-channel requantize
could load them with a single vld1q_s32.

Switch storage to int8 and load via vldrbq_s32 (load 4 int8s,
sign-extend to int32) at the same latency.  Scalar paths read
shifts[oc] and let the C type system sign-extend on assignment to
int32 — bit-exact with the prior int32 storage.

Memory: .rodata for full MV2 drops 51,168 B (from 4,284,212 to
4,233,044, ~1.2%).  Bigger relative win for narrower models with
similar channel counts.

Latency: end-to-end MV2 on Corstone-300 FVP drops slightly from
48,975,234 to 47,964,461 cycles (1.02x) — the smaller array reduces
cache pressure but the vldrbq vs vld1q has the same throughput on
M55.

Bit-exact preserved (max int8 diff = 0 vs cortex_m runtime reference)
across all four phase tests.

Authored with assistance from Claude (claude.ai/code).

+++ examples/models/mv2_cortex_m_mve/include/mv2_layer_params.h
@@ -60,7 +60,7 @@ typedef struct {
   const int8_t* weight;            // [out_c, kernel_h, kernel_w, in_c] (OHWI)
   const int32_t* bias;             // [out_c]; nullable
   const int32_t* requant_mults;    // [out_c]
-  const int32_t* requant_shifts;   // [out_c]
+  const int8_t*  requant_shifts;   // [out_c] — fits in int8 (frexp exponent, ~[-30, +5])
   int32_t input_offset;
   int32_t output_offset;
   int8_t activation_min;
@@ -84,7 +84,7 @@ typedef struct {
   const int8_t* weight;            // [1, kernel_h, kernel_w, out_c] (IHWO)
   const int32_t* bias;             // [out_c]; nullable
   const int32_t* requant_mults;    // [out_c]
-  const int32_t* requant_shifts;   // [out_c]
+  const int8_t*  requant_shifts;   // [out_c] — int8 packed
   int32_t input_offset;
   int32_t output_offset;
   int8_t activation_min;
